+++ tests/compat/rhino-adapter.mjs
@@ -1,9 +1,3 @@
-const [version] = argv._;
+const [path] = argv._;
 
-const rhino = await fetch(
-  `https://github.com/mozilla/rhino/releases/download/Rhino${ version.replace(/\./g, '_') }_Release/rhino-${ version }.jar`,
-);
-
-await fs.writeFile('tests/compat/rhino.jar', Buffer.from(await rhino.arrayBuffer()));
-
-await $`java -jar tests/compat/rhino.jar -version 200 -require tests/compat/rhino-runner.js`;
+await $`java -jar ${ path } -version 200 -require tests/compat/rhino-runner.js`;